v.1.3.1 overhaul updates
took out the better trims mod and replaced it with something more compatible.. ( hard to get silver items when the mod doesnt add silver ore or require a mod that does.. ) dont fret, the combination of two mods i put in make it infinitely more diverse and useful. I added more to the nether and the deep dark as well as the deeper darker dimension as well. should keep people moderately busy.